backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.30
pydantic-settings==2.2.1
pydantic==2.7.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-multipart==0.0.9
google-generativeai==0.9.0
httpx==0.27.0

